Fixed login

+++ Sources/MapleStoryServer/Migrations/CreateWorld.swift
@@ -12,14 +12,14 @@ struct CreateWorld: AsyncMigration {
     func prepare(on database: Database) async throws {
         try await database.schema(World.schema)
             .id()
-            .field(World.CodingKeys.index, .uint8, .required)
+            .field(World.CodingKeys.index, .int, .required)
             .field(World.CodingKeys.name, .string, .required)
             .field(World.CodingKeys.address, .string, .required)
-            .field(World.CodingKeys.flags, .uint8, .required)
+            .field(World.CodingKeys.flags, .int, .required)
             .field(World.CodingKeys.eventMessage, .string, .required)
-            .field(World.CodingKeys.rateModifier, .uint8, .required)
-            .field(World.CodingKeys.eventXP, .uint8, .required)
-            .field(World.CodingKeys.dropRate, .uint8, .required)
+            .field(World.CodingKeys.rateModifier, .int, .required)
+            .field(World.CodingKeys.eventXP, .int, .required)
+            .field(World.CodingKeys.dropRate, .int, .required)
             .create()
     }
